Update styles and components to support dark mode
