+++ sents-web/src/app/landing/page.tsx
@@ -1,70 +1,15 @@
-'use client';
-import React, { useEffect } from 'react';
 import Link from 'next/link';
 import LandingIllustration from '@/public/images/landingIllustration.png';
 import LandingImage from '@/public/images/landingImage.png';
 import LandingImage2 from '@/public/images/landingImage2.png';
 import Image from 'next/image';
-import { BsFillMoonStarsFill } from 'react-icons/bs';
-import { MdOutlineKeyboardArrowDown, MdOutlineLightMode } from 'react-icons/md';
 import { GoArrowRight } from 'react-icons/go';
-import { TfiWorld } from 'react-icons/tfi';
-import { useTheme } from 'next-themes';
+import LandingHeader from '@/components/header/Landing';
 
 const LandingPage = () => {
-  const { theme, setTheme } = useTheme();
-
-  useEffect(() => {
-    document.body.classList.toggle('dark', theme === 'dark');
-  }, [theme]);
-
-  const toggleTheme = () => {
-    setTheme(theme === 'dark' ? 'light' : 'dark');
-  };
-
   return (
     <div className="container relative mx-auto px-4 h-full">
-      <header className="flex justify-between flex-wrap items-center pt-6">
-        <h1 className="text-3xl Unigoe-font tracking-wider text-[#0D4222] dark:text-[#E6F6F0]">
-          SENTS.
-        </h1>
-        <nav className="max-sm:flex max-sm:justify-center max-sm:w-full">
-          <ul className="flex items-center dark:text-white h-auto space-x-6 py-4 px-5 bg-[#E6F6F0] dark:bg-[#39463E80] rounded-full">
-            <li>
-              <Link
-                href="/login_register"
-                className="p-3 rounded-xl bg-white dark:bg-[#39463E80]"
-              >
-                Sign in
-              </Link>
-            </li>
-            <li>
-              <Link
-                href="/login_register"
-                className="p-3 rounded-xl bg-[#1EF1A5] dark:text-[#0D4222]"
-              >
-                Sign Up
-              </Link>
-            </li>
-            <li>
-              <button className="mt-1" onClick={toggleTheme}>
-                {theme === 'dark' ? (
-                  <MdOutlineLightMode size={20} />
-                ) : (
-                  <BsFillMoonStarsFill size={20} />
-                )}
-              </button>
-            </li>
-            <li>
-              <button className="flex items-center">
-                <TfiWorld className="text-[#0D4222] dark:text-white" />
-                <MdOutlineKeyboardArrowDown className="text-[#0D4222] dark:text-white" />
-              </button>
-            </li>
-          </ul>
-        </nav>
-      </header>
-
+      <LandingHeader />
       <main className="relative text-center flex flex-col justify-between  h-full">
         <section className="relative text-[#0D4222] dark:text-[#E6F6F0] py-12 z-50">
           <h2 className="text-3xl leading-[65px] md:text-[80px] py-6 Unigoe-font">
@@ -101,9 +46,15 @@ const LandingPage = () => {
 
         <section>
           <Image
-            src={theme === 'dark' ? LandingImage2 : LandingImage}
+            src={LandingImage}
             alt="Landing Image"
-            className="object-contain w-full relative bottom-[125px] lg:bottom-[14rem] 2xl:bottom-[18rem]"
+            className="object-contain w-full relative bottom-[125px] lg:bottom-[14rem] 2xl:bottom-[18rem] dark:hidden"
+            loading="eager"
+          />
+          <Image
+            src={LandingImage2}
+            alt="Landing Image"
+            className="object-contain w-full relative bottom-[125px] lg:bottom-[14rem] 2xl:bottom-[18rem] hidden dark:block"
             loading="eager"
           />
         </section>
