+++ sents-web/src/app/layout.tsx
@@ -1,6 +1,7 @@
 import '@/styles/styles.scss';
 import Provider from '@/services/redux/Provider';
 import ThemeProvider from '../components/themeProvider/ThemeProvider';
+import ThemeToggle from '@/components/themeProvider/ThemeToggle';
 
 export const metadata = {
   title: 'Next.js',
@@ -22,7 +23,8 @@ export default function RootLayout({
             enableSystem
             disableTransitionOnChange
           >
-            {children}
+            <ThemeToggle />
+            <>{children}</>
           </ThemeProvider>
         </Provider>
       </body>
