+++ sents-web/src/assets/styles/styles.scss
@@ -32,6 +32,6 @@ html {
   font-family: 'Unigeo64';
 }
 
-body.dark {
+.dark {
   background-color: #0e120f;
 }
